simplifies code for finding & loading Dynatrace dashboards
